+++ src/kpdvdb/__init__.py
@@ -1,6 +1,6 @@
 """KayPENTAX Disordered Voice Database Reader module"""
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 
 import pandas as pd
 from os import path