+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - remove_unknowns xtor argument to drop entries with missing info
 
+### Fixed
+
+- diagnoses DF to exclude NA entries
+
+
 ## [0.3.1] - 2023-01-17
 
 ### Fixed